requirements.txt
flask
webargs
click
requests
nh-currency